fix: refactor webhook tests and implementation to separate private methods into new classes
